Added Support Directory + Files
Added Support directory to help with the installation and setting up of argus and soon to be zeek. I also added a README file inside the Support directory so it can explain what the directories are and what the files inside them are. This file is still under development, unfinished. I need to add the rest of my files, the argus.conf, argusarchive and so on. For another day.