Integrated to Capstone_Flask
*WARNING not all files, models, datasets are included because they are just too big*

Updated app.py, end user is now able to use capture.py, will just need to update the IP to the main servers one to upload it. And if on windows will either need tshark added to path or to run it in administrative prompt. There are terminal options that you can do to choose how long, the name, and the interface there are default options to all of these. Still just through CLI. Need to work on preprocessing and predicting with models then to return the result.