Major Update
Implemented SQLite to handle user account creation and storing results for Network_Classifier. The capture.exe's are still **IN PROGRESS**. There is going to be one for Windows and Mac that'll try to use Tshark first, if not it'll fallback on their respective's native network capture tool, pktmon (Win), or TCPDUMP (macOS). And the XGB and DT joblib have been excluded because they are too big. They can be found in our One Drive Directory.